Restructure packaging for wechat
COAND-497
