+++ wechatpay/src/main/java/com/adyen/checkout/wechatpay/internal/ui/WeChatDelegate.kt
@@ -6,13 +6,15 @@
  * Created by oscars on 19/8/2022.
  */
 
-package com.adyen.checkout.wechatpay
+package com.adyen.checkout.wechatpay.internal.ui
 
+import androidx.annotation.RestrictTo
 import com.adyen.checkout.components.base.ActionDelegate
 import com.adyen.checkout.components.base.DetailsEmittingDelegate
 import com.adyen.checkout.components.base.IntentHandlingDelegate
 import com.adyen.checkout.components.ui.ViewProvidingDelegate
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 interface WeChatDelegate :
     ActionDelegate,
     DetailsEmittingDelegate,
